+++ src/core/mod.rs
@@ -1,3 +1,4 @@
+mod handle_message;
 mod kex;
 mod metacommands;
 
@@ -10,17 +11,14 @@ use message::OxyMessage::{self, *};
 use pty::Pty;
 use shlex;
 use std::{
-    cell::RefCell, collections::HashMap, fs::{metadata, symlink_metadata, File}, io::{Read, Write}, net::ToSocketAddrs, path::PathBuf, rc::Rc,
-    time::{Duration, Instant},
+    cell::RefCell, collections::HashMap, fs::File, io::Read, path::PathBuf, rc::Rc, time::{Duration, Instant},
 };
 use transportation::{
-    self, mio::{
-        net::{TcpListener, TcpStream}, PollOpt, Ready, Token,
-    }, set_timeout, BufferedTransport, EncryptedTransport,
-    EncryptionPerspective::{Alice, Bob}, MessageTransport, Notifiable, Notifies, ProtocolTransport,
+    self, mio::net::TcpListener, set_timeout, BufferedTransport, EncryptedTransport, EncryptionPerspective::{Alice, Bob}, MessageTransport,
+    Notifiable, Notifies, ProtocolTransport,
 };
 #[cfg(unix)]
-use tuntap::{TunTap, TunTapType};
+use tuntap::TunTap;
 use ui::Ui;
 
 #[derive(Clone)]
@@ -87,8 +85,12 @@ impl Oxy {
             #[cfg(unix)]
             tuntaps: Rc::new(RefCell::new(HashMap::new())),
         };
-        let proxy = NakedNotificationProxy { oxy: x.clone() };
-        x.naked_transport.borrow_mut().as_mut().unwrap().set_notify(Rc::new(proxy));
+        let proxy = x.clone();
+        x.naked_transport
+            .borrow_mut()
+            .as_mut()
+            .unwrap()
+            .set_notify(Rc::new(move || proxy.notify_naked()));
         x
     }
 
@@ -97,8 +99,8 @@ impl Oxy {
             return;
         }
         *self.ui.borrow_mut() = Some(Ui::create());
-        let proxy = UiNotificationProxy { oxy: self.clone() };
-        let proxy = Rc::new(proxy);
+        let proxy = self.clone();
+        let proxy = Rc::new(move || proxy.notify_ui());
         self.ui.borrow_mut().as_ref().unwrap().set_notify(proxy);
     }
 
@@ -141,257 +143,6 @@ impl Oxy {
         trace!("Sending message {}: {:?}", message_number, message);
         self.underlying_transport.borrow().as_ref().unwrap().send(message);
         message_number
-    }
-
-    fn handle_message(&self, message: OxyMessage, message_number: u64) {
-        debug!("Recieved message {}", message_number);
-        trace!("Received message {}: {:?}", message_number, message);
-        match message {
-            DummyMessage { .. } => (),
-            Ping {} => {
-                self.send(Pong {});
-            }
-            Pong {} => {
-                *self.last_message_seen.borrow_mut() = Instant::now();
-            }
-            BasicCommand { command } => {
-                assert!(perspective() == Bob);
-                #[cfg(unix)]
-                let sh = "/bin/sh";
-                #[cfg(unix)]
-                let flag = "-c";
-                #[cfg(windows)]
-                let sh = "cmd.exe";
-                #[cfg(windows)]
-                let flag = "/c";
-                let result = ::std::process::Command::new(sh).arg(flag).arg(command).output();
-                if let Ok(result) = result {
-                    self.send(BasicCommandOutput {
-                        stdout: result.stdout,
-                        stderr: result.stderr,
-                    });
-                }
-            }
-            #[cfg(unix)]
-            PtyRequest { command } => {
-                assert!(perspective() == Bob);
-                let pty = Pty::forkpty(&command);
-                pty.underlying.set_notify(Rc::new(PtyNotificationProxy { oxy: self.clone() }));
-                *self.pty.borrow_mut() = Some(pty);
-                self.send(PtyRequestResponse { granted: true });
-                trace!("Successfully allocated PTY");
-            }
-            #[cfg(unix)]
-            PtyRequestResponse { granted } => {
-                assert!(perspective() == Alice);
-                if !granted {
-                    warn!("PTY open failed");
-                    return;
-                }
-                let (w, h) = self.ui.borrow_mut().as_mut().unwrap().pty_size();
-                self.send(PtySizeAdvertisement { w, h });
-            }
-            #[cfg(unix)]
-            PtySizeAdvertisement { w, h } => {
-                assert!(perspective() == Bob);
-                self.pty.borrow_mut().as_mut().unwrap().set_size(w, h);
-            }
-            #[cfg(unix)]
-            PtyInput { data } => {
-                assert!(perspective() == Bob);
-                if self.pty.borrow_mut().is_none() {
-                    self.send(Reject {
-                        message_number,
-                        note: "No PTY exists".to_string(),
-                    });
-                    return;
-                }
-                self.pty.borrow_mut().as_mut().unwrap().underlying.put(&data[..]);
-            }
-            #[cfg(unix)]
-            PtyOutput { data } => {
-                assert!(perspective() == Alice);
-                self.ui.borrow_mut().as_mut().unwrap().pty_data(&data);
-            }
-            BasicCommandOutput { stdout, stderr } => {
-                assert!(perspective() == Alice);
-                info!("BasicCommandOutput {:?}, {:?}", stdout, stderr);
-                if let Ok(stdout) = String::from_utf8(stdout) {
-                    info!("stdout:\n-----\n{}\n-----", stdout);
-                }
-            }
-            DownloadRequest { path } => {
-                assert!(perspective() == Bob);
-                let file = File::open(path);
-                if file.is_err() {
-                    self.send(Reject {
-                        message_number,
-                        note: "Failed to open file".to_string(),
-                    });
-                    return;
-                }
-                let file = file.unwrap();
-                let metadata = file.metadata();
-                if metadata.is_err() {
-                    self.send(Reject {
-                        message_number,
-                        note: "Failed to stat file".to_string(),
-                    });
-                    return;
-                }
-                let metadata = metadata.unwrap();
-                self.send(FileSize {
-                    reference: message_number,
-                    size:      metadata.len(),
-                });
-                self.transfers_out.borrow_mut().push((message_number, file));
-            }
-            UploadRequest { path, filepart } => {
-                assert!(perspective() == Bob);
-                if let Ok(meta) = metadata(&path) {
-                    if meta.is_dir() {
-                        let mut buf: PathBuf = path.into();
-                        buf.push(filepart);
-                        let file = File::create(buf).unwrap();
-                        self.transfers_in.borrow_mut().insert(message_number, file);
-                        return;
-                    }
-                }
-                let file = File::create(path).unwrap();
-                self.transfers_in.borrow_mut().insert(message_number, file);
-            }
-            FileSize { reference: _, size } => {
-                ::copy::push_file_size(size);
-            }
-            FileData { reference, data } => {
-                if data.is_empty() {
-                    debug!("File transfer completed");
-                    self.transfers_in.borrow_mut().remove(&reference);
-                    if self.copy_peer.borrow_mut().is_some() {
-                        *self.fetch_file_ticker.borrow_mut() += 1;
-                        if *self.fetch_file_ticker.borrow_mut() < arg::matches().occurrences_of("source") {
-                            let filename = arg::matches()
-                                .values_of("source")
-                                .unwrap()
-                                .nth(*self.fetch_file_ticker.borrow_mut() as usize)
-                                .unwrap();
-                            let filename = filename.splitn(2, ':').nth(1).unwrap().to_string();
-                            let cmd = ["download", &filename, "unused"].to_vec();
-                            let cmd = cmd.iter().map(|x| x.to_string()).collect();
-                            self.handle_metacommand(cmd);
-                        } else {
-                            let mut message = [0u8; 8];
-                            byteorder::BE::write_u64(&mut message, ::std::u64::MAX);
-                            self.copy_peer.borrow_mut().as_ref().unwrap().send_message(&message);
-                            trace!("Source is done.");
-                        }
-                    }
-                    return;
-                }
-                if self.copy_peer.borrow_mut().is_none() {
-                    self.transfers_in.borrow_mut().get_mut(&reference).unwrap().write_all(&data[..]).unwrap();
-                } else {
-                    let ticker = *self.fetch_file_ticker.borrow_mut();
-                    let mut message: Vec<u8> = Vec::new();
-                    message.resize(8, 0);
-                    byteorder::BE::write_u64(&mut message[..8], ticker);
-                    message.extend(&data);
-                    self.copy_peer.borrow_mut().as_ref().unwrap().send_message(&message);
-                }
-            }
-            BindConnectionAccepted { reference } => {
-                assert!(perspective() == Alice);
-                let addr = self.remote_bind_destinations.borrow_mut().get(&reference).unwrap().parse().unwrap();
-                let stream = TcpStream::connect(&addr).unwrap();
-                let bt = BufferedTransport::from(stream);
-                let stream = PortStream {
-                    stream: bt,
-                    token:  message_number,
-                    oxy:    self.clone(),
-                    local:  false,
-                };
-                let stream2 = Rc::new(stream.clone());
-                stream.stream.set_notify(stream2);
-                self.remote_streams.borrow_mut().insert(message_number, stream);
-            }
-            RemoteOpen { addr } => {
-                assert!(perspective() == Bob);
-                let dest = addr.to_socket_addrs().unwrap().next().unwrap();
-                debug!("Resolved RemoteOpen destination to {:?}", dest);
-                let stream = TcpStream::connect(&dest).unwrap();
-                let bt = BufferedTransport::from(stream);
-                let stream = PortStream {
-                    stream: bt,
-                    token:  message_number,
-                    oxy:    self.clone(),
-                    local:  false,
-                };
-                let stream2 = Rc::new(stream.clone());
-                stream.stream.set_notify(stream2);
-                self.remote_streams.borrow_mut().insert(message_number, stream);
-            }
-            RemoteBind { addr } => {
-                assert!(perspective() == Bob);
-                let bind = TcpListener::bind(&addr.parse().unwrap()).unwrap();
-                let proxy = BindNotificationProxy {
-                    oxy:   self.clone(),
-                    token: Rc::new(RefCell::new(message_number)),
-                };
-                let proxy = Rc::new(proxy);
-                let token = transportation::insert_listener(proxy.clone());
-                transportation::borrow_poll(|poll| {
-                    poll.register(&bind, Token(token), Ready::readable(), PollOpt::level()).unwrap();
-                });
-                let bind = PortBind {
-                    listener:    bind,
-                    local_spec:  addr,
-                    remote_spec: "".to_string(),
-                };
-                self.port_binds.borrow_mut().insert(message_number, bind);
-            }
-            RemoteStreamData { reference, data } => {
-                self.remote_streams.borrow_mut().get_mut(&reference).unwrap().stream.put(&data[..]);
-            }
-            LocalStreamData { reference, data } => {
-                self.local_streams.borrow_mut().get_mut(&reference).unwrap().stream.put(&data[..]);
-            }
-            #[cfg(unix)]
-            TunnelRequest { tap, name } => {
-                self.bob_only();
-                let mode = if tap { TunTapType::Tap } else { TunTapType::Tun };
-                let tuntap = TunTap::create(mode, &name, message_number, self.clone());
-                self.tuntaps.borrow_mut().insert(message_number, tuntap);
-            }
-            #[cfg(unix)]
-            TunnelData { reference, data } => {
-                let borrow = self.tuntaps.borrow_mut();
-                borrow.get(&reference).unwrap().send(&data);
-            }
-            StatRequest { path } => {
-                let info = symlink_metadata(path);
-                if info.is_err() {
-                    self.send(Reject {
-                        message_number,
-                        note: "stat failed".to_string(),
-                    });
-                    return;
-                }
-                let info = info.unwrap();
-                let message = StatResult {
-                    reference:         message_number,
-                    len:               info.len(),
-                    is_dir:            info.is_dir(),
-                    is_file:           info.is_file(),
-                    atime:             None,
-                    ctime:             None,
-                    mtime:             None,
-                    owner:             "".to_string(),
-                    group:             "".to_string(),
-                    octal_permissions: 0,
-                };
-            }
-            _ => (),
-        }
     }
 
     #[cfg(unix)]
@@ -519,15 +270,16 @@ impl Oxy {
         key.extend(keys::static_key());
         let et = EncryptedTransport::create(bt, arg::perspective(), &key);
         let pt = ProtocolTransport::create(et);
-        pt.set_notify(Rc::new(self.clone()));
+        let proxy = self.clone();
+        pt.set_notify(Rc::new(move || proxy.notify_main_transport()));
         *self.underlying_transport.borrow_mut() = Some(pt);
-        self.notify();
+        self.notify_main_transport();
         self.do_post_auth();
     }
 
     fn register_signal_handler(&self) {
-        let proxy = SignalNotificationProxy { oxy: self.clone() };
-        transportation::set_signal_handler(Rc::new(proxy));
+        let proxy = self.clone();
+        transportation::set_signal_handler(Rc::new(move || proxy.notify_signal()));
     }
 
     fn notify_signal(&self) {
@@ -551,7 +303,8 @@ impl Oxy {
                 let cmd = cmd.iter().map(|x| x.to_string()).collect();
                 self.handle_metacommand(cmd);
             }
-            let proxy = TransferNotificationProxy { oxy: self.clone() };
+            let proxy = self.clone();
+            let proxy = move || proxy.notify_transfer();
             self.copy_peer.borrow_mut().as_mut().unwrap().set_notify(Rc::new(proxy));
             self.notify_transfer();
             return;
@@ -567,7 +320,8 @@ impl Oxy {
             self.create_ui();
         }
         self.register_signal_handler();
-        set_timeout(Rc::new(KeepaliveProxy { oxy: self.clone() }), Duration::from_secs(60));
+        let proxy = self.clone();
+        set_timeout(Rc::new(move || proxy.notify_keepalive()), Duration::from_secs(60));
     }
 
     fn run_batched_metacommands(&self) {
@@ -634,7 +388,8 @@ impl Oxy {
             ::std::process::exit(2);
         }
         self.send(Ping {});
-        set_timeout(Rc::new(KeepaliveProxy { oxy: self.clone() }), Duration::from_secs(60));
+        let proxy = self.clone();
+        set_timeout(Rc::new(move || proxy.notify_keepalive()), Duration::from_secs(60));
     }
 
     fn notify_socks_bind(&self, token: u64) {
@@ -703,10 +458,8 @@ impl Oxy {
             }
         }
     }
-}
 
-impl Notifiable for Oxy {
-    fn notify(&self) {
+    fn notify_main_transport(&self) {
         trace!("Core notified");
         if self.underlying_transport.borrow().as_ref().unwrap().is_closed() {
             #[cfg(unix)]
@@ -744,27 +497,6 @@ impl Notifiable for Oxy {
         }
         self.service_transfers();
         self.notify_transfer(); // Uhh... this is a function naming disaster. REFACTOR
-    }
-}
-
-struct UiNotificationProxy {
-    oxy: Oxy,
-}
-
-impl Notifiable for UiNotificationProxy {
-    fn notify(&self) {
-        self.oxy.notify_ui();
-    }
-}
-
-struct PtyNotificationProxy {
-    oxy: Oxy,
-}
-
-impl Notifiable for PtyNotificationProxy {
-    fn notify(&self) {
-        #[cfg(unix)]
-        self.oxy.notify_pty();
     }
 }
 
@@ -818,16 +550,6 @@ impl Notifiable for PortStream {
     }
 }
 
-struct NakedNotificationProxy {
-    oxy: Oxy,
-}
-
-impl Notifiable for NakedNotificationProxy {
-    fn notify(&self) {
-        self.oxy.notify_naked();
-    }
-}
-
 struct SocksConnectionNotificationProxy {
     oxy:   Oxy,
     bt:    BufferedTransport,
@@ -844,34 +566,4 @@ impl Notifiable for SocksConnectionNotificationProxy {
 enum SocksState {
     Initial,
     Authed,
-}
-
-struct TransferNotificationProxy {
-    oxy: Oxy,
-}
-
-impl Notifiable for TransferNotificationProxy {
-    fn notify(&self) {
-        self.oxy.notify_transfer();
-    }
-}
-
-struct SignalNotificationProxy {
-    oxy: Oxy,
-}
-
-impl Notifiable for SignalNotificationProxy {
-    fn notify(&self) {
-        self.oxy.notify_signal();
-    }
-}
-
-struct KeepaliveProxy {
-    oxy: Oxy,
-}
-
-impl Notifiable for KeepaliveProxy {
-    fn notify(&self) {
-        self.oxy.notify_keepalive();
-    }
 }
